Rewrite the README with an eye towards first-time contributors. (#96)
This commit radically overhauls the README for the Documentation website
by providing a more structured introduction to the process of editing a
Hugo website. It moves more general information (such as how to operate
a Git client) into external references. These references are linked, as
places such as the official GitHub documentation will do a better job of
maintaining "how to create a pull request" documentation than we will.

Also added was a table of contents in order to more easily navigate the
expanded readme, a far more detailed description of prerequisites needed
to make edits to the documentation, and numerous readability
enhancements.