First pass at major reorgnization.
This is a big commit. I have radically reorganized the structure of the
Docs website in order to continue the work that began in #97. I have
changed the names and filepaths of numerous documentation sections so
that they are ready to receive "guidebook"-style improvements. Notably,
this includes the "Networking" and "Organization" sections, whose names
I found ambiguous and whose contents could be better distinguished, IMO.

The new names of the documentation sections should (I hope) make clear
the direction in which I am working: I want it to be possible for a
complete NYC Mesh "newbie" to look at the outline of the documentation
generated by Hugo and have a clear idea of where they should start
reading to accomplish a given task.

This is aided by a complete rewrite of the "Getting Started" page, which
now offers additional guidance about which part of the documentation to
read based on the role or goal that the reader wants to accomplish.

Additionally, I have expanded the outline by adding numerous new
sections, including an expanded "Continuing Education" section (which
now houses the links to NYC Mesh's "Networking Classes" along with links
to other free and by-donation in-person computer networking workshops
organized by NYC Mesh volunteers and contributors), a new
as-yet-unwritten "Troublehsooting Guide," and an important layperson's
"Overview" section that describes technical details about the NYC Mesh
implementation.

Using Hugo's `aliases` feature, I have also supplied numerous redirects
to various important parts of the old documentation for which the URL
permalinks have changed, so that these links do not break. However, this
URL redirection coverage may not be complete as I am not intimately
familiar with all the places that incoming links may be published.

This is a work-in-progress, but I believe is complete enough for the
direction of my work to be clear.