3/8" Hex Socket -> 7mm Nut Driver
At least for the Omni/LBE 25-70mm hose clamps that I have on hand, it seems that a 7mm socket is a much better fit than 5/16" socket. Feel free to reject if other installers disagree.

Also, although less descriptive, it seems that “Nut Driver bit” is a better search term than “Hex Socket bit” when looking for the product online.